Refactor code and replace all function key to using arrow functions
