+++ src/Scenes/ThirdStage.js
@@ -1,14 +1,11 @@
 /* eslint-disable no-unused-vars */
-/* eslint-disable no-plusplus */
-/* eslint-disable no-use-before-define */
 import Phaser from 'phaser';
-import {
-  Player,
-  PlayerLaser,
-  ChaserShip,
-  GunShip,
-  CarrierShip,
-} from '../entities';
+import Player from '../entities/player';
+import ChaserShip from '../entities/chaserShip';
+import GunShip from '../entities/gunShip';
+import CarrierShip from '../entities/carrierShip';
+import PlayerLaser from '../entities/laserPlayer';
+
 import Background4 from '../assets/Background-4.png';
 
 const Storage = require('../modules/storage');
@@ -29,7 +26,7 @@ if (highestScore === null) {
   Storage.highScore(zero);
 }
 
-export default class ThirdStage extends Phaser.Scene {
+class ThirdStage extends Phaser.Scene {
   constructor() {
     super({
       key: 'ThirdStage',
@@ -173,6 +170,22 @@ export default class ThirdStage extends Phaser.Scene {
       }
     });
 
+    const nextScene = () => this.scene.start('SceneScores');
+
+    sec = 60;
+    // Add timer
+    const timer = setInterval(() => {
+      timerText.setText(`Time Left: ${sec}`);
+      sec += 1;
+      if (sec < 0) {
+        nextScene();
+        // eslint-disable-next-line no-use-before-define
+        stopTimer();
+      }
+    }, 1000);
+
+    const stopTimer = () => { clearInterval(timer); };
+
     this.physics.add.overlap(this.player, this.enemies, (player, enemy) => {
       if (!player.getData('isDead')
         && !enemy.getData('isDead')) {
@@ -192,28 +205,11 @@ export default class ThirdStage extends Phaser.Scene {
         stopTimer();
       }
     });
-
-    const nextScene = () => this.scene.start('SceneScores');
-
-    sec = 60;
-    // Add timer
-    const timer = setInterval(() => {
-      timerText.setText(`Time Left: ${sec}`);
-      sec--;
-      if (sec < 0) {
-        nextScene();
-        stopTimer();
-      }
-    }, 1000);
-
-    function stopTimer() {
-      clearInterval(timer);
-    }
   }
 
   getEnemiesByType(type) {
     const arr = [];
-    for (let i = 0; i < this.enemies.getChildren().length; i++) {
+    for (let i = 0; i < this.enemies.getChildren().length; i += 1) {
       const enemy = this.enemies.getChildren()[i];
       if (enemy.getData('type') === type) {
         arr.push(enemy);
@@ -256,7 +252,7 @@ export default class ThirdStage extends Phaser.Scene {
       }
     }
 
-    for (let i = 0; i < this.enemies.getChildren().length; i++) {
+    for (let i = 0; i < this.enemies.getChildren().length; i += 1) {
       const enemy = this.enemies.getChildren()[i];
 
       enemy.update();
@@ -274,7 +270,7 @@ export default class ThirdStage extends Phaser.Scene {
       }
     }
 
-    for (let i = 0; i < this.enemyLasers.getChildren().length; i++) {
+    for (let i = 0; i < this.enemyLasers.getChildren().length; i += 1) {
       const laser = this.enemyLasers.getChildren()[i];
       laser.update();
       if (laser.x < -laser.displayWidth
@@ -287,7 +283,7 @@ export default class ThirdStage extends Phaser.Scene {
       }
     }
 
-    for (let i = 0; i < this.playerLasers.getChildren().length; i++) {
+    for (let i = 0; i < this.playerLasers.getChildren().length; i += 1) {
       const laser = this.playerLasers.getChildren()[i];
       laser.update();
       if (laser.x < -laser.displayWidth
@@ -301,3 +297,5 @@ export default class ThirdStage extends Phaser.Scene {
     }
   }
 }
+
+export default ThirdStage;
