+++ src/entities/enemyLaser.js
@@ -1,0 +1,11 @@
+import Entity from './entity';
+import 'phaser';
+
+class EnemyLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprLaserEnemy0');
+    this.body.velocity.y = 200;
+  }
+}
+
+export default EnemyLaser;
