+++ src/entities/gunShip.js
@@ -1,0 +1,37 @@
+import Phaser from 'phaser';
+import Entity from './entity';
+import EnemyLaser from './enemyLaser';
+
+class GunShip extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprEnemy0', 'GunShip');
+    this.play('sprEnemy0');
+
+    this.body.velocity.y = Phaser.Math.Between(50, 100);
+
+    this.shootTimer = this.scene.time.addEvent({
+      delay: 1000,
+      callback() {
+        const laser = new EnemyLaser(
+          this.scene,
+          this.x,
+          this.y,
+        );
+        laser.setScale(this.scaleX);
+        this.scene.enemyLasers.add(laser);
+      },
+      callbackScope: this,
+      loop: true,
+    });
+  }
+
+  onDestroy() {
+    if (this.shootTimer !== undefined) {
+      if (this.shootTimer) {
+        this.shootTimer.remove(false);
+      }
+    }
+  }
+}
+
+export default GunShip;
