+++ src/entities/laserPlayer.js
@@ -1,0 +1,11 @@
+import Entity from './entity';
+import 'phaser';
+
+class PlayerLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'sprLaserPlayer');
+    this.body.velocity.y = -200;
+  }
+}
+
+export default PlayerLaser;
