+++ src/entities/player.js
@@ -1,0 +1,76 @@
+import Phaser from 'phaser';
+import Entity from './entity';
+import PlayerLaser from './laserPlayer';
+
+let ammunition = 100;
+const Storage = require('../modules/storage');
+
+class Player extends Entity {
+  constructor(scene, x, y, key) {
+    super(scene, x, y, key, 'Player');
+    this.setData('speed', 200);
+    this.setData('isShooting', false);
+    this.setData('timerShootDelay', 10);
+    this.setData('timerShootTick', this.getData('timerShootDelay') - 1);
+  }
+
+  moveUp() {
+    this.body.velocity.y = -this.getData('speed');
+  }
+
+  moveDown() {
+    this.body.velocity.y = this.getData('speed');
+  }
+
+  moveLeft() {
+    this.body.velocity.x = -this.getData('speed');
+  }
+
+  moveRight() {
+    this.body.velocity.x = this.getData('speed');
+  }
+
+  onDestroy() {
+    this.scene.time.addEvent({
+      delay: 1000,
+      callback() {
+        this.scene.scene.start('SceneScores');
+      },
+      callbackScope: this,
+      loop: false,
+    });
+  }
+
+  onSecondStage() {
+    this.scene.time.addEvent({
+      delay: 1000,
+      callback() {
+        this.scene.scene.start('SceneSecondStage');
+      },
+      callbackScope: this,
+      loop: false,
+    });
+  }
+
+  update() {
+    this.body.setVelocity(0, 0);
+    this.x = Phaser.Math.Clamp(this.x, 0, this.scene.game.config.width);
+    this.y = Phaser.Math.Clamp(this.y, 0, this.scene.game.config.height);
+
+    if (this.getData('isShooting')) {
+      if (this.getData('timerShootTick') < this.getData('timerShootDelay')) {
+        this.setData('timerShootTick', this.getData('timerShootTick') + 1);
+      } else {
+        const laser = new PlayerLaser(this.scene, this.x, this.y);
+        this.scene.playerLasers.add(laser);
+
+        this.scene.sfx.laser.play();
+        this.setData('timerShootTick', 0);
+        ammunition -= 1;
+        Storage.setAmmo(ammunition);
+      }
+    }
+  }
+}
+
+export default Player;
