+++ src/modules/getScore.js
@@ -1,5 +1,5 @@
 const GetScore = (() => {
-  async function all() {
+  const all = async () => {
     const response = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/ZUi2Xo2RRfSKd14twwPn/scores/', {
       method: 'GET',
       headers: {
@@ -9,7 +9,7 @@ const GetScore = (() => {
     });
     const data = await response.json();
     return data;
-  }
+  };
   return {
     all,
   };
