+++ src/modules/storage.js
@@ -1,30 +1,30 @@
 const Storage = (() => {
-  function currentScore(score) {
+  const currentScore = (score) => {
     localStorage.setItem('currentScore', JSON.stringify(score));
-  }
+  };
 
-  function getCurrentScore() {
+  const getCurrentScore = () => {
     const getScore = JSON.parse(localStorage.getItem('currentScore'));
     return getScore;
-  }
+  };
 
-  function highScore(score) {
+  const highScore = (score) => {
     localStorage.setItem('highestScore', JSON.stringify(score));
-  }
+  };
 
-  function getHighScore() {
+  const getHighScore = () => {
     const getHigh = JSON.parse(localStorage.getItem('highestScore'));
     return getHigh;
-  }
+  };
 
-  function setAmmo(ammo) {
+  const setAmmo = (ammo) => {
     localStorage.setItem('Ammunition', JSON.stringify(ammo));
-  }
+  };
 
-  function currentAmmo() {
+  const currentAmmo = () => {
     const ammo = JSON.parse(localStorage.getItem('Ammunition'));
     return ammo;
-  }
+  };
   return {
     currentScore,
     getCurrentScore,
