+++ src/modules/submitScore.js
@@ -1,5 +1,5 @@
 const SubmitScore = (() => {
-  async function send(name, score) {
+  const send = async (name, score) => {
     const response = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/ZUi2Xo2RRfSKd14twwPn/scores/', {
       method: 'POST',
       headers: {
@@ -13,7 +13,7 @@ const SubmitScore = (() => {
     });
     const data = await response.json();
     return data;
-  }
+  };
   return {
     send,
   };
