+++ src/tests/submitScore.test.js
@@ -1,0 +1,8 @@
+const SubmitScore = require('../modules/submitScore');
+
+it('submits the player score to the leaderboard', () => {
+  SubmitScore.send('PlayerOne', 0)
+    .then((response) => {
+      expect(response).toBe('Leaderboard score created correctly.');
+    });
+});
